utils: Move filesystem helpers to submodule
